Cleaned Pytorch Folder
